+++ src/components/Icon.tsx
@@ -103,7 +103,7 @@ function IconMusic(props: SVGProps<SVGSVGElement>): ReactNode {
 }
 
 // https://icons.getbootstrap.com/icons/dash-circle/
-function IconOpen(props: SVGProps<SVGSVGElement>): ReactNode {
+function IconMinus(props: SVGProps<SVGSVGElement>): ReactNode {
   return (
     <svg
       xmlns="http://www.w3.org/2000/svg"
@@ -119,7 +119,7 @@ function IconOpen(props: SVGProps<SVGSVGElement>): ReactNode {
 }
 
 // https://icons.getbootstrap.com/icons/dash-circle/
-function IconClosed(props: SVGProps<SVGSVGElement>): ReactNode {
+function IconPlus(props: SVGProps<SVGSVGElement>): ReactNode {
   return (
     <svg
       xmlns="http://www.w3.org/2000/svg"
@@ -165,10 +165,46 @@ function IconSparkles(props: SVGProps<SVGSVGElement>): ReactNode {
   );
 }
 
+// https://icons.getbootstrap.com/icons/list/
+function IconMenuHamburger(props: SVGProps<SVGSVGElement>): ReactNode {
+  return (
+    <svg
+      xmlns="http://www.w3.org/2000/svg"
+      width="16"
+      height="16"
+      fill="currentColor"
+      viewBox="0 0 16 16"
+      {...props}
+    >
+      <path
+        fillRule="evenodd"
+        d="M2.5 12a.5.5 0 0 1 .5-.5h10a.5.5 0 0 1 0 1H3a.5.5 0 0 1-.5-.5m0-4a.5.5 0 0 1 .5-.5h10a.5.5 0 0 1 0 1H3a.5.5 0 0 1-.5-.5m0-4a.5.5 0 0 1 .5-.5h10a.5.5 0 0 1 0 1H3a.5.5 0 0 1-.5-.5"
+      />
+    </svg>
+  );
+}
+
+// https://icons.getbootstrap.com/icons/x-lg/
+function IconClose(props: SVGProps<SVGSVGElement>): ReactNode {
+  return (
+    <svg
+      xmlns="http://www.w3.org/2000/svg"
+      width="16"
+      height="16"
+      fill="currentColor"
+      viewBox="0 0 16 16"
+      {...props}
+    >
+      <path d="M2.146 2.854a.5.5 0 1 1 .708-.708L8 7.293l5.146-5.147a.5.5 0 0 1 .708.708L8.707 8l5.147 5.146a.5.5 0 0 1-.708.708L8 8.707l-5.146 5.147a.5.5 0 0 1-.708-.708L7.293 8z" />
+    </svg>
+  );
+}
+
 const icons = {
-  open: IconOpen,
-  closed: IconClosed,
+  minus: IconMinus,
+  plus: IconPlus,
   clear: IconClear,
+  close: IconClose,
   music: IconMusic,
   search: IconSearch,
   sparkles: IconSparkles,
@@ -176,6 +212,7 @@ const icons = {
   arrowRight: IconArrowRight,
   arrowLeftDouble: IconArrowLeftDouble,
   arrowRightDouble: IconArrowRightDouble,
+  menuHamburger: IconMenuHamburger,
 } as const;
 
 export interface IconProps {