+++ src/hooks/useScrollToFragment.ts
@@ -2,7 +2,8 @@ import { useLayoutEffect } from "react";
 import { useLocation } from "react-router-dom";
 
 export function useScrollToFragment(): void {
-  const fragment = useLocation().hash.slice(1);
+  const location = useLocation();
+  const fragment = location.hash.slice(1);
 
   useLayoutEffect(() => {
     if (!fragment) {