+++ src/screens/ScreenOffense.tsx
@@ -1,8 +1,9 @@
 import { clsx } from "clsx";
 import { ReactNode, useEffect } from "react";
 import { useTranslation } from "react-i18next";
-import { useNavigate } from "react-router-dom";
+import { useLocation, useNavigate } from "react-router-dom";
 import { useSessionStorage } from "usehooks-ts";
+import { Card } from "../components/Card";
 import {
   CheckboxGroup,
   CheckboxGroupOption,
@@ -27,8 +28,6 @@ import {
   typesFromString,
 } from "../misc/data-types";
 import styles from "./ScreenOffense.module.css";
-import { Card } from "../components/Card";
-import { OffenseTabs } from "../components/OffenseTabs";
 
 export type ScreenOffenseProps = {
   mode: "combination" | "single";
@@ -43,6 +42,7 @@ export function ScreenOffense({ mode }: ScreenOffenseProps): ReactNode {
   const { t, i18n } = useTranslation();
   const search = useSearch();
   const navigate = useNavigate();
+  const location = useLocation();
   const [offenseTypes, setOffenseTypes] = useSessionStorage<Type[]>(
     "offense.types",
     [],
@@ -62,8 +62,8 @@ export function ScreenOffense({ mode }: ScreenOffenseProps): ReactNode {
     if (search.has("abilities")) {
       setAbilities(splitTokens(search.get("abilities") || "") as any);
     }
-    navigate({ search: "" }, { replace: true });
-  }, [search]);
+    navigate({ search: "", hash: location.hash }, { replace: true });
+  }, [search, location.hash]);
 
   useEffect(() => {
     setOffenseTypes((offenseTypes) =>
@@ -126,10 +126,6 @@ export function ScreenOffense({ mode }: ScreenOffenseProps): ReactNode {
 
   return (
     <main className={clsx(styles.root, "content-wide center")}>
-      <div className={styles.tabBar}>
-        <OffenseTabs />
-      </div>
-
       <Flex direction="column" gap="xlarge">
         <Flex direction="column" gap="medium">
           <FancyText tag="h2" fontSize="large" fontWeight="medium">